by attheshow, tyleryoungblood, edmunddunn, COBadger, pixelwhip: Adds Sprint 39 changes (please see release notes).

+++ web/modules/contrib/date_recur/src/DateRecurOccurrences.php
@@ -44,28 +44,28 @@ class DateRecurOccurrences implements EventSubscriberInterface, EntityTypeListen
    *
    * @var \Drupal\Core\Database\Connection
    */
-  protected $database;
+  protected Connection $database;
 
   /**
    * The entity type manager.
    *
    * @var \Drupal\Core\Entity\EntityTypeManagerInterface
    */
-  protected $entityTypeManager;
+  protected EntityTypeManagerInterface $entityTypeManager;
 
   /**
    * The entity field manager.
    *
    * @var \Drupal\Core\Entity\EntityFieldManagerInterface
    */
-  protected $entityFieldManager;
+  protected EntityFieldManagerInterface $entityFieldManager;
 
   /**
    * Manages data type plugins.
    *
    * @var \Drupal\Core\TypedData\TypedDataManagerInterface
    */
-  protected $typedDataManager;
+  protected TypedDataManagerInterface $typedDataManager;
 
   /**
    * DateRecurOccurrences constructor.
@@ -265,6 +265,10 @@ class DateRecurOccurrences implements EventSubscriberInterface, EntityTypeListen
    */
   protected function fieldStorageDelete(FieldStorageDefinitionInterface $fieldDefinition): void {
     $tableName = static::getOccurrenceCacheStorageTableName($fieldDefinition);
+    if (!$this->database->schema()->tableExists($tableName)) {
+      return;
+    }
+
     $this->database
       ->schema()
       ->dropTable($tableName);
@@ -421,14 +425,12 @@ class DateRecurOccurrences implements EventSubscriberInterface, EntityTypeListen
    */
   protected function getBaseFieldStorages(ContentEntityTypeInterface $entityType): array {
     $baseFields = $this->entityFieldManager->getBaseFieldDefinitions($entityType->id());
-    $baseFields = array_filter($baseFields, function (FieldDefinitionInterface $fieldDefinition): bool {
-      return $this->isDateRecur($fieldDefinition->getFieldStorageDefinition());
-    });
+    $baseFields = array_filter($baseFields,
+      fn (FieldDefinitionInterface $fieldDefinition): bool => $this->isDateRecur($fieldDefinition->getFieldStorageDefinition())
+    );
 
     return array_map(
-      function (FieldDefinitionInterface $baseField): FieldStorageDefinitionInterface {
-        return $baseField->getFieldStorageDefinition();
-      },
+      fn (FieldDefinitionInterface $baseField): FieldStorageDefinitionInterface => $baseField->getFieldStorageDefinition(),
       $baseFields
     );
   }
